+++ .github/helpers/count_molecule_matrix.py
@@ -37,6 +37,7 @@ def main(event_name, repo_owner, review_state, ref):
     if event_name == 'workflow_dispatch' or review_state == 'approved' or ref == 'refs/heads/master':
         ce_matrix.append(get_ce_params(molecule_scenario='update_cartridge'))
         ce_matrix.append(get_ce_params(molecule_scenario='check_facts'))
+        ce_matrix.append(get_ce_params(molecule_scenario='rolling_update'))
         ce_matrix.append(get_ce_params(tarantool_version='1.10'))
         ce_matrix.append(get_ce_params(ansible_version='2.8.0'))
         ce_matrix.append(get_ce_params(ansible_version='2.9.0'))